+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.WebApi/Program.cs
@@ -24,7 +24,7 @@ namespace GreenEnergyHub.Charges.WebApi
             CreateHostBuilder(args).Build().Run();
         }
 
-        private static IHostBuilder CreateHostBuilder(string[] args) =>
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
             Host.CreateDefaultBuilder(args)
                 .ConfigureWebHostDefaults(webBuilder =>
                 {